[F-3][PE-9] [Create API that could allow user to CRUD a PSE stock tickers
