+++ plutus_eye/extensions.py
@@ -1,0 +1,5 @@
+from flask_pymongo import PyMongo
+from flask_jwt_extended import JWTManager
+
+mongo = PyMongo()
+jwt = JWTManager()
