+++ plutus_eye/settings.py
@@ -1,6 +1,5 @@
 import os
 
-
 MONGO_URI = os.environ.get('MONGO_URI')
 JWT_SECRET_KEY = os.environ.get('JWT_SECRET_KEY')
 GATEWAY_TOKEN = os.environ.get('GATEWAY_TOKEN')
