+++ plutus_eye/utils/ticker.py
@@ -5,6 +5,10 @@ health_care = ['JNJ', 'MRK', 'PFE', 'UNH', 'AMGN', 'ACB', 'SMG', 'CARA', 'ZYNE',
 commodities = ['AA', 'CVX', 'DD', 'XOM', 'TX', 'PBR', 'VALE', 'NBL', 'MOS', 'RDS.B', 'FMC', 'DOW', 'COP', 'ET', 'EPD', 'PAGP', 'INT', 'HAL', 'IP', 'NOV', 'FCX', 'NUE', 'HES', 'CHK', 'PBF', 'HFC', 'DVN', 'EOG', 'X', 'KMI', 'PPG', 'APA', 'HUN', 'SHW', 'RS', 'APD', 'NGL', 'EMN', 'STLD', 'STI', 'TRGP', 'DK', 'MUR', 'FTI', 'ASH', 'NEM', 'CMC', 'CE', 'BTU', 'MRC', 'SE', 'CLMT', 'PXD', 'SRLP', 'SON', 'DNO', 'FOE', 'PRS', 'TGS', 'SAND', 'EVK', 'BNR', 'RCP', 'DLN', 'AXTA', 'DNOW', 'KRO', 'LNG', 'NWPX', 'RIG', 'SAEX', 'SU', 'RDS.A', 'BHP', 'RIO', 'LIN', 'PBR.A', 'ENB', 'SLB', 'OXY', 'LYB', 'CXO', 'FANG', 'IFF', 'MRO', 'COG', 'XEC', 'PE', 'WLL']
 finance = ["AXP", "BAC", "JPM", "TRV", "C", "HBC", "MA", "V", "DB", "QIWI", "SAN", "BRK.B", "WU", "DEM", "AGNC", "ALV", "MUV2", "DB1", "CBK", "ACA", "BNP", "CS", "GLE", "UL", "AZM", "BMPS", "BP", "BPE", "EXO", "GASI", "ISP", "MDBI", "MED", "PMI", "UBI", "UCG", "US", "BBVA", "BKIA", "BKT", "BME", "CABK", "MAP", "POP", "SAB", "SBERL", "CSGN", "BAER", "CFR", "SLHN", "SREN", "UBSG", "ZURN", "COF", "GS", "MET", "WFC", "FNMA", "AIG", "ALL", "AFL", "USB", "PGR", "HIG", "BK", "PNC", "L", "CB", "LNC", "MMC", "JEF", "AMP", "BLK", "VOYA", "RGA", "STT", "UNM", "PFG", "AIZ", "BBT", "ALLY", "DFS", "GNW", "FNF", "CBRE", "BEN", "KKR", "BX", "WRB", "FITB", "SCHW", "AFG", "NAVI", "ORI", "JLL", "RF", "HST", "RLGY", "Y", "MKL", "THG", "RJF", "CINF", "SPG", "Z", "MTB", "FAF", "AJG", "ICE", "NTRS", "LPLA", "KEY", "CNO", "AMT", "AFSI", "TROW", "TMK", "ARGO", "MS", "ADM", "ADN", "AV", "BARC", "BLND", "HL", "HMSO", "HSBA", "LAND", "LGEN", "LLOY", "LSE", "OML", "PRU", "RBS", "RSA", "RSL", "SDR", "SLA", "STAN", "III", "BPM", "SGRO.L", "SMT.L", "QLT.L", "DNB", "NDA_SE", "SEB-A", "SHB-A", "SWED-A", "INVE-A", "KINV-A", "INDU-A", "LUND-B", "ORSTED", "TRYG", "NDA1V", "SAMPO", "0939.HK", "3988.HK", "1398.HK", "1288.HK", "2007.HK", "0998.HK", "1988.HK", "1339.HK", "1359.HK", "2318.HK", "2328.HK", "2628.HK", "3328.HK", "0688.HK", "1299.HK", "2601.HK", "3968.HK", "1966.HK", "6030.HK", "1628.HK", "0656.HK", "1109.HK", "00011.HK", "00016.HK", "00017.HK", "00388.HK", "00966.HK", "01113.HK", "01336.HK", "02388.HK", "03333.HK", "03908.HK", "06060.HK", "06837.HK", "06881.HK", "06886.HK", "DWNI", "WDI", "GYC", "TEG", "AOX", "ARL", "PAT", "DEQ", "VNA.DE", "AT1.DE", "LEG.DE", "DWS.DE", "GLJ.DE", "TLG.DE", "ADJ.DE", "MGAM", "COA", "SHI", "HAS", "SMDS", "HTG", "JLG", "RSW.L", "BBA", "SOPH", "STOB", "ISAT", "UBM", "IWG", "TMPL", "PAGE", "SGC", "ASCL", "CEY", "PNL", "HOC", "ACA.L", "JAM", "JLT", "HICL", "FDM", "CINE", "BKG", "TRIG", "EVR", "IBST", "EMG", "MLC", "VSVS", "GNK", "UKW", "RPC", "INCH", "DOM", "MONY", "QQ", "BEZ", "ASL", "INVP", "CTY", "BNKR", "SAFE", "SMWH", "WKP", "HSV", "PNN", "BWNG", "CLDN", "MNKS", "FCIT", "DCG", "BME.L", "ECM", "AGR", "3IN", "NRR", "VCT", "BTEM", "RDI", "UDG", "GPOR", "GNC.L", "MAB", "MARS", "BAG", "JUST", "CARD", "JDW", "SNN", "KIE", "CBG", "JLIF", "DNLM", "BVS", "ASHM", "GNS", "NOG", "BTG.L", "VEIL", "JII", "SXX", "VEC", "PSH", "JE", "RSE", "VOF", "SEQI", "SDP", "PFG.L", "PHP", "PIN", "JEO", "FCS", "EFM", "CCFS", "CGEO", "BGFD", "LQDE", "IAEX", "IAPD", "SWDA", "IEML", "ISF", "IMIB", "SEMB", "IFFF", "IBZL", "IDEM", "IMEU", "IJPE", "ITWN", "IUSA", "IUKP", "VXX", "CPJ1", "CSPX", "CUKX", "EMIM", "ERNE", "GHYS", "IEUX", "IGLS", "IGLT", "IGUS", "IJPH", "INXG", "IWDP", "MIDD", "SGLN", "SJPA", "SLXX", "XIV", "SHY", "IEF", "DBA", "CORN", "WEAT", "UGA", "INDY", "BKF", "IAI", "RWR", "SGG", "BAL", "JO", "PBW", "PPA", "EWY", "ILF", "EWM", "EWW", "EWT", "GGOV", "LQD", "BOND", "GSY", "FPE", "HYLS", "WIP", "IGOV", "IHY", "IBND", "AUNZ", "HYEM", "SPFF", "FTSL", "VXXB", "ACWI", "AMG", "BPY", "BSAC", "CIT", "CME", "DHIL", "ESGR", "ESNT", "FXCM", "GLPI", "GRC", "IBKR", "LAMR", "MAA", "SBNY", "SEIC", "SUPV", "LM", "LAZ", "CACC", "RNR", "MMI", "FCPT", "RMAX", "BGN.MI", "BPM.MI", "FBK.MI", "VTI", "VOO", "ITUB", "VEA", "BBD", "VWO", "CCI", "VTV", "PLD", "IWF", "PRU", "VUG", "EQIX", "IWD", "PSA", "BND", "VNQ", "WELL", "AMTD", "AVB", "EQR", "DLR", "USMV", "VCSH", "BSV", "VEU", "SYF", "IVW", "VCIT", "VTR", "O", "TIP", "BXP", "VGT", "IWB", "SHV", "IWR", "FRC", "EMB", "CFG", "PFF", "NLY", "HBAN", "VGK", "IAU", "SIVB", "CMA", "ETFC", "MINT", "QUAL", "SPLV", "JNK", "ADS", "IWO", "AMLP", "IWN", "ZION", "MTUM", "BIL", "EZU", "XLC", "BKLN", "INDA", "SLV", "XBI", "MCHI", "AAXJ", "BHF", "XLB", "ELLI", "EWH", "XLRE", "KRE", "FEZ", "EWG", "ASHR", "KBE", "KWEB", "XS6R.L", "IBGM.L", "IGLO.L", "CRPS.L", "SQQQ", "TVIX", "TZA", "FAZ", "SOXX", "SCO", "XCX5.L", "ABN.NV", "AGN.NV", "ASRNL.NV", "FLOW.NV", "HAL.NV", "INTER.NV", "NIBC.NV", "NN.NV", "PSH.NV", "TFG.NV", "WHA.NV", "AMUN.PA", "CNP.PA", "ENX.PA", "LI.PA", "KN.PA", "NXI.PA", "PLUS", "CMCX", "GCAP"]
 services = ["AMZN", "DIS", "HD", "MCD", "WMT", "TWX", "EBAY", "DTV", "TV", "LVS", "GPS", "URBN", "NFLX", "ANF", "FDX", "CAR", "SBUX", "PETM", "BABA", "ETSY", "RYAAY", "OSTK", "LYFT", "UBER", "BYND", "DPW", "LHA", "AC", "AIR", "ORA", "PUB", "ABE", "TL5", "CTRP", "JOBS", "ADEN", "ACN", "AAL", "BKNG", "SYY", "IM", "BBY", "DAL", "UAL", "SHLD", "TJX", "TECD", "AVT", "RAD", "TWC", "ARW", "SPLS", "MAN", "KSS", "LUV", "SVU", "AES", "GLP", "ODP", "GPC", "OMC", "MUSA", "ARMK", "DISH", "WFM", "SNX", "VIAB", "JWN", "CHRW", "YUM", "KMX", "JEC", "CSX", "JCP", "RRD", "BBBY", "LB", "ROST", "HLT", "HSIC", "MGM", "GWW", "GPI", "AAP", "AZO", "OMI", "GME", "SAH", "HDS", "CHTR", "DRI", "CZR", "NWSA", "ACM", "CORE", "SPTN", "WCC", "TA", "IPG", "ORLY", "CASY", "LYV", "DKS", "UNFI", "DDS", "R", "EXPD", "CVC", "AXE", "BKS", "KBR", "DISCA", "JBHT", "ABG", "JBLU", "EXPE", "URI", "KELYA", "BAH", "WYNN", "ALK", "ESND", "WYND", "SPGI", "BIG", "YRCW", "FISV", "ABM", "QUAD", "BURL", "MIK", "RUSHB", "WSM", "RHI", "CTAS", "FWONA", "BLMN", "H", "VWR", "KNX", "TIF", "CMG", "PDCO", "WSO", "GNC", "MAR", "SQ", "MTCH", "HRMS", "AGK", "AHT", "BAB", "BNZL", "BRBY", "SKY", "CPG", "CPI", "EXPN", "EZJ", "GFS", "IAG", "IHG", "ITRK", "ITV", "KGF", "MKS", "MRW", "NXT", "PSON", "REL", "RMG", "SBRY", "TSCO", "WMH", "WPP", "WTB", "BDEV.L", "DCC.L", "DLG.L", "INF.L", "MCRO.L", "PPB.L", "RTO.L", "SKG.L", "TUI.L", "ASC.L", "RYA.L", "GJF", "GOGL", "NAS", "STB", "TELIA", "0762.HK", "0728.HK", "0700.HK", "00001.HK", "00019.HK", "00763.HK", "03690.HK", "O2D", "KGX", "FNTN", "UTDI", "RRTL", "HNR1.DE", "TLX.DE", "FRA.DE", "DHER.DE", "G24.DE", "SPR.DE", "SIX2.DE", "PSM.DE", "SAX.DE", "PCT", "CYBG", "ALLE", "AN", "ARII", "CHDN", "UNIT", "DPZ", "EAT", "FLT", "GHC", "HAIN", "HRI", "HTZ", "LABL", "LBTYA", "LBTYK", "LGF", "LILA", "LSXMA", "MCO", "MEG", "MELI", "WBT", "MMYT", "MNST", "P", "QSR", "SAVE", "SCI", "SERV", "SIG", "TIVO", "VIPS", "VRSK", "W", "WLTW", "LEE", "SIX", "AON", "OUT", "PST.MI", "REC.MI", "UNI.MI", "TMUS", "SIRI", "ANET", "IAC", "ATUS", "OKTA", "FFIV", "TTD", "ZS", "ROKU", "ZAYO", "GRUB", "CIEN", "IGHY.L", "FTI.PA", "BOL.PA", "IPS.PA", "MMB.PA", "MMT.PA", "FSLY", "FVRR"]
+basic_materials = ["AA", "CVX", "DD", "XOM", "TX", "PBR", "VALE", "POT", "NBL", "MOS", "BAS", "LNA", "TKA", "SDF", "LXSd", "FP", "MTS", "SOLB", "TEC", "ENI", "SPM", "TEN", "ABG-P", "LKODL", "MNODL", "ROSNL", "SVSTL", "GIVN", "LHN", "SYNN", "UHR", "RDS.B", "FMC", "DOW", "COP", "ET", "EPD", "PAGP", "INT", "ANDV", "HAL", "IP", "BHI", "NOV", "FCX", "NUE", "HES", "CHK", "PBF", "HFC", "DVN", "APC", "EOG", "X", "KMI", "MON", "PPG", "WNR", "APA", "PX", "HUN", "SHW", "CAM", "RS", "APD", "NGL", "EMN", "STLD", "STI", "TRGP", "DK", "MUR", "FTI", "ASH", "NEM", "CMC", "CE", "BTU", "BPL", "AKS", "MRC", "SE", "CLMT", "PXD", "NTI", "SRLP", "SON", "LINE", "BMS", "CLR", "CF", "SPN", "CLF", "AXLL", "VAL", "WLK", "RPM", "AE", "SWN", "CEQP", "ARG", "AAL", "AMFW", "ANTO", "BG", "BHP", "BP", "FRES", "GLEN", "JMAT", "PFC", "RDSA", "RIO", "RRS", "TLW", "CRDA.L", "AKA", "AKSO", "DNO", "FOE", "PGS", "PRS", "SUBC", "TGS", "VOLV-A", "SAND", "BOL", "LUPE", "VWS", "SSABAH", "STEAV", "UPM", "NESTE", "0386.HK", "0883.HK", "1088.HK", "1171.HK", "3800.HK", "2899.HK", "1898.HK", "0968.HK", "2883.HK", "00857.HK", "03993.HK", "NDX1", "SGCG", "EVK", "LXS", "1COV", "KCO", "GXI", "SZG", "BNR", "SY1", "RAA.DE", "NDA.DE", "VOE.DE", "FPE3.DE", "KWS.DE", "RCP", "IGG", "DRX", "DLN", "RTN.L", "SVS", "LMP", "TALK", "BWY", "SAGA", "FGT", "SRP", "HVPE", "OCDO", "RHIM", "RWI", "FENR", "GOLD.", "AXTA", "CVI.N", "CVRR", "DNOW", "KRO", "LNG", "NWPX", "RIG", "SAEX", "SU", "WFT", "RDS.A", "BHP", "RIO", "LIN", "PBR.A", "DWDP", "ENB", "SLB", "OXY", "LYB", "BHGE", "CXO", "FANG", "IFF", "MRO", "COG", "ECA", "GG", "XEC", "PE", "VSM", "WLL", "AKZA.NV", "AMG.NV", "DSM.NV", "IMCD.NV", "OCI.NV""IMCD.NV", "OCI.NV"]
+conglomerates = ["MMM", "IEP", "SEB", "SPB"]
+industrial_goods= ["BA", "CAT", "GE", "UTX", "SI", "CX", "LMT", "SIE", "HEI", "ALO", "DG", "EN", "LG", "LR", "SAF", "SGO", "SU", "ATL", "BZU", "CNHI", "LDO", "PRY", "STS", "ACS", "FCC", "FER", "SGRE", "OHL", "SCYR", "TRE", "PHORL", "URKAL", "ABB", "SGSN", "HON", "UNP", "UPS", "DE", "GD", "EMR", "NOC", "RTN", "FLR", "DHR", "CMI", "ITW", "WM", "TXT", "PH", "LLL", "NSC", "SWK", "NAV", "AGCO", "PCP", "RSG", "MAS", "DOV", "DHI", "WY", "PWR", "LEN", "HII", "SPR", "ROK", "EME", "BGC", "PHM", "OC", "COL", "FLS", "OA", "SPW", "MDU", "CFX", "MTZ", "TPC", "NVR", "ATI", "BEAV", "AME", "DAR", "XYL", "TOL", "MTW", "SEDG", "NTLA", "EDIT", "CLLS", "BAl", "CRH", "IMI", "MGGT", "MRO", "PSN", "RR", "SMIN", "TPK", "WEIR", "FERG", "TW.L", "NHY", "YAR", "SKF-A", "ASSA-B", "ATCO-A", "ALFA", "SECU-B", "SKA-B", "TREL-B", "DSV", "ISS", "KNEBV", "WRT1V", "3323.HK", "1800.HK", "0390.HK", "0914.HK", "1618.HK", "1186.HK", "00285.HK", "00788.HK", "02018.HK", "02382.HK", "HDD", "G1A", "GBF", "HLAG", "OSR", "HOT.DE", "WCH.DE", "KU2.DE", "JUN3.DE", "HWDN", "SNR", "TCAP", "COB", "HSTG", "HSTN", "CRST", "POLY", "ELM", "PHNX", "AUTO", "ESUR", "PLP", "VM", "HLMA", "PETS", "GRI", "CSP", "AA.L", "KAZ", "ABBN", "DCI", "FTV", "IR", "LAYN", "MWA", "ROP", "SRCL", "STRL", "AOBC", "TDG", "TPH", "USG", "BRE.MI", "ETN", "KEYS", "VMC", "MLM", "WAB", "KSU", "WRK", "OLED", "XPO", "AALB.NV", "ARCAD.NV", "BOKA.NV", "PNL.NV", "LIGHT.NV", "TWEKA.NV", "KVW.NV", "ALT.PA", "AM.PA", "FGR.PA", "GET.PA", "ADP.PA", "TKTT.PA", "HO.PA", "VK.PA"]
+utilities = ["PEGI", "CVA", "EVA", "NEP", "ALB", "ENS", "EOAN", "RWE", "EDF", "ENGI", "VIE", "A2A", "EGPW", "ENEL", "SRG", "TLIT", "TRN", "ANA", "ELE", "ENG", "NTGY", "IBE", "REE", "REP", "OGZDL", "KMGL", "KCELL", "MFONL", "NVTKL", "SGGDL", "ATADL", "PSX", "WMB", "VLO", "MPC", "EXC", "DUK", "SO", "PCG", "NEE", "AEP", "NRG", "FE", "EIX", "ED", "OKE", "ETR", "D", "XEL", "PPL", "SRE", "PEG", "CNP", "UGI", "CPN", "ES", "CMS", "NI", "AEE", "WEC", "SCG", "ATO", "POM", "CNA", "NG", "SSE", "SVT", "UU", "TLN.L", "BWLPG", "TEL", "DANSKE", "FORTUM", "ELISA", "0902.HK", "1816.HK", "0916.HK", "00002.HK", "00006.HK", "00135.HK", "00257.HK", "00384.HK", "02386.HK", "02688.HK", "UN01", "LEO", "DRI", "IGY.DE", "DJAN", "SPX", "RMV", "CCC", "FCPT", "DC", "UKCM", "SMP", "GCP", "FCSS", "MRC", "BRW", "TED", "NMC", "GVC", "FEV", "JMG", "PDL", "FGP", "GLO", "CAA", "CACQ", "MSEX", "PAM", "YORW", "IG.MI", "DTE", "AWK", "EVRG", "VST", "LNT", "ATC.NV", "SBMO.NV", "WKL.NV", "ILD.PA", "SEV.PA", "TCH.PA", "TFI.PA", "PEGI", "CVA", "EVA", "NEP", "ALB", "ENS", "EOAN", "RWE", "EDF", "ENGI", "VIE", "A2A", "EGPW", "ENEL", "SRG", "TLIT", "TRN", "ANA", "ELE", "ENG", "NTGY", "IBE", "REE", "REP", "OGZDL", "KMGL", "KCELL", "MFONL", "NVTKL", "SGGDL", "ATADL", "PSX", "WMB", "VLO", "MPC", "EXC", "DUK", "SO", "PCG", "NEE", "AEP", "NRG", "FE", "EIX", "ED", "OKE", "ETR", "D", "XEL", "PPL", "SRE", "PEG", "CNP", "UGI", "CPN", "ES", "CMS", "NI", "AEE", "WEC", "SCG", "ATO", "POM", "CNA", "NG", "SSE", "SVT", "UU", "TLN.L", "BWLPG", "TEL", "DANSKE", "FORTUM", "ELISA", "0902.HK", "1816.HK", "0916.HK", "00002.HK", "00006.HK", "00135.HK", "00257.HK", "00384.HK", "02386.HK", "02688.HK", "UN01", "LEO", "DRI", "IGY.DE", "DJAN", "SPX", "RMV", "CCC", "FCPT", "DC", "UKCM", "SMP", "GCP", "FCSS", "MRC", "BRW", "TED", "NMC", "GVC", "FEV", "JMG", "PDL", "FGP", "GLO", "CAA", "CACQ", "MSEX", "PAM", "YORW", "IG.MI", "DTE", "AWK", "EVRG", "VST", "LNT", "ATC.NV", "SBMO.NV", "WKL.NV", "ILD.PA", "SEV.PA", "TCH.PA", "TFI.PA"]
 
 default_tickers = {
     'technology': tech,
@@ -12,5 +16,9 @@ default_tickers = {
     'health_care': health_care,
     'commodities': commodities,
     'finance': finance,
-    'services': services
+    'services': services,
+    'basic_materials': basic_materials,
+    'conglomerates': conglomerates,
+    'industrial_goods': industrial_goods,
+    'utilities': utilities
 }
